+++ tools/gulp-tasks/docs-app.js
@@ -1,0 +1,3 @@
+module.exports = (gulp) => () => {
+  gulp.src('docs/src/**/*').pipe(gulp.dest('dist/docs'));
+};
